requirements.txt
Django==2.0.2
mysqlclient==1.3.12
pillow==5.0.0
gunicorn
djangorestframework==3.7.7
